# swagger.yaml
openapi: 3.0.0
info:
  title: Todo List API
  version: 1.0.0
  description: A comprehensive Todo List API with user authentication and CRUD operations
  contact:
    name: API Support
    email: support@todoapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    User:
      type: object
      required:
        - username
        - email
        - password
      properties:
        _id:
          type: string
          description: The auto-generated id of the user
        username:
          type: string
          description: The username of the user
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          description: The email of the user
        password:
          type: string
          description: The password of the user
          minLength: 6
        createdAt:
          type: string
          format: date-time
          description: The date the user was created
        updatedAt:
          type: string
          format: date-time
          description: The date the user was last updated
    Todo:
      type: object
      required:
        - title
      properties:
        _id:
          type: string
          description: The auto-generated id of the todo
        title:
          type: string
          description: The title of the todo
          maxLength: 100
        description:
          type: string
          description: The description of the todo
          maxLength: 500
        completed:
          type: boolean
          description: The completion status of the todo
          default: false
        priority:
          type: string
          enum:
            - low
            - medium
            - high
          description: The priority level of the todo
          default: medium
        dueDate:
          type: string
          format: date
          description: The due date of the todo
        user:
          type: string
          description: The ID of the user who owns the todo
        createdAt:
          type: string
          format: date-time
          description: The date the todo was created
        updatedAt:
          type: string
          format: date-time
          description: The date the todo was last updated
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT token for authentication
    TodoResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            todo:
              $ref: '#/components/schemas/Todo'
    TodoListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            todos:
              type: array
              items:
                $ref: '#/components/schemas/Todo'
            total:
              type: integer
              description: Total number of todos
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Number of todos per page
            pages:
              type: integer
              description: Total number of pages
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: Detailed error information
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Health check endpoint
      tags:
        - Health
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  version:
                    type: string
  /auth/signup:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  description: Username (3-30 characters)
                  minLength: 3
                  maxLength: 30
                email:
                  type: string
                  format: email
                  description: User email address
                password:
                  type: string
                  minLength: 6
                  description: User password (minimum 6 characters)
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User email address
                password:
                  type: string
                  description: User password
      responses:
        '200':
          description: User successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /todos:
    get:
      summary: Get all todos for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of todos per page
        - in: query
          name: completed
          schema:
            type: boolean
          description: Filter by completion status
        - in: query
          name: priority
          schema:
            type: string
            enum:
              - low
              - medium
              - high
          description: Filter by priority level
      responses:
        '200':
          description: List of todos retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Todo title
                  maxLength: 100
                description:
                  type: string
                  description: Todo description
                  maxLength: 500
                priority:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                  description: Todo priority level
                dueDate:
                  type: string
                  format: date
                  description: Todo due date
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /todos/{id}:
    get:
      summary: Get a specific todo by ID
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '200':
          description: Todo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Todo ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Todo title
                  maxLength: 100
                description:
                  type: string
                  description: Todo description
                  maxLength: 500
                completed:
                  type: boolean
                  description: Todo completion status
                priority:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                  description: Todo priority level
                dueDate:
                  type: string
                  format: date
                  description: Todo due date
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User authentication endpoints
  - name: Todos
    description: Todo management endpoints